backend/requirements.txt
# FastAPI and web framework - Latest versions
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database and ORM - Latest versions
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
asyncpg==0.29.0

# Modern AI/ML Libraries
openai==1.3.7
anthropic==0.7.7
langchain==0.0.350
langchain-openai==0.0.2
langchain-anthropic==0.0.1
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
accelerate==0.24.1
bitsandbytes==0.41.3
peft==0.6.0
trl==0.7.4

# Advanced Data Science
numpy==1.24.3
pandas==2.1.3
polars==0.19.19
scikit-learn==1.3.2
scipy==1.11.4
plotly==5.17.0
streamlit==1.28.1

# Financial Analysis - Enhanced
yfinance==0.2.28
pandas-ta==0.3.14b0
ta==0.10.2
ccxt==4.1.77
alpha-vantage==2.3.1
quandl==3.7.0

# Modern Web Technologies
websockets==12.0
aiofiles==23.2.1
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Data Validation and Settings
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic-extra-types==2.4.0

# Authentication and Security - Enhanced
bcrypt==4.1.2
cryptography==41.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Caching and Performance
redis==5.0.1
aioredis==2.0.1
cachetools==5.3.2

# Monitoring and Observability
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Testing and Development - Enhanced
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
jinja2==3.1.2
click==8.1.7
rich==13.7.0
typer==0.9.0
